TASK: refactor steps to separate components
